Split ToopherIframe into own file
